requirements.txt
selenium==4.11.2
pytest==7.4.2
allure-pytest==2.13.2